pyproject.toml
[project]
name = "tmp-ddd-modular-monolith"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.2",
    "asyncpg>=0.30.0",
    "dishka>=1.5.3",
    "faststream[rabbit]>=0.5.43",
    "litestar[jwt]>=2.16.0",
    "loguru>=0.7.3",
    "passlib[argon2]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=0.26.0",
    "python-jose>=3.4.0",
    "sqlalchemy[asyncio]>=2.0.40",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.1",
]

[tool.ruff]
line-length = 120